Two changes:
FIRST: Any Get/Post variables that begin with "hld_" are caught and loaded to the object, so they will be in $this->hld, an associative array.  Also, any variables in that array will be passed out to the page as hidden variables.  This lets you carry around context automatically from page to page.

SECOND: Detail pane generation now calls projection() in androLib.php to generate inputs.  This makes custom pages (which must use projection) consistent with default pages.