Added page x6phpinfo, so developers and admins always have access to this useful information on any running site.
To use it, go to ?x6page=phpinfo.  This works on all current versions of Andromeda, even though we are not publishing any docs on x6 yet.  

This program *requires* you to be a root user on the node manager.  It is not possible for regular system administrators to grant acess to the program.